socialapi: implemented better concurrency and moved folder structure
