--- go/src/socialapi/workers/slackapi/oauth/api.go
@@ -1,243 +1,0 @@
-package api
-
-import (
-	"fmt"
-	kodingmodels "koding/db/models"
-	"koding/db/mongodb/modelhelper"
-	"net/http"
-	"net/url"
-	"socialapi/models"
-	"socialapi/workers/common/response"
-	"sync"
-	"time"
-
-	"github.com/nlopes/slack"
-
-	"golang.org/x/oauth2"
-	"gopkg.in/mgo.v2"
-	"gopkg.in/mgo.v2/bson"
-)
-
-const slackAway = "away"
-
-type Slack struct {
-	OAuthConf *oauth2.Config
-}
-
-//
-// Send & Callback handlers are used for slack oauth
-// If we can get the token successfully, then we store user's token in mongo in Callback handler
-//
-
-func (s *Slack) Send(w http.ResponseWriter, req *http.Request) {
-
-	url := s.OAuthConf.AuthCodeURL("state", oauth2.AccessTypeOffline)
-	http.Redirect(w, req, url, http.StatusTemporaryRedirect)
-}
-
-func (s *Slack) Callback(u *url.URL, h http.Header, _ interface{}, c *models.Context) (int, http.Header, interface{}, error) {
-	// state := req.FormValue("state")
-	// if state != oauthStateString {
-	// 	fmt.Printf("invalid oauth state, expected '%s', got '%s'\n", oauthStateString, state)
-	// 	http.Redirect(w, req, "/", http.StatusTemporaryRedirect)
-	// 	return
-	// }
-
-	code := u.Query().Get("code")
-
-	token, err := s.OAuthConf.Exchange(oauth2.NoContext, code)
-	if err != nil {
-		fmt.Printf("oauthConf.Exchange() failed with '%s'\n", err)
-		return response.NewBadRequest(err)
-	}
-
-	user, err := modelhelper.GetUser(c.Client.Account.Nick)
-	if err != nil && err != mgo.ErrNotFound {
-		return response.NewBadRequest(err)
-	}
-
-	if err == mgo.ErrNotFound {
-		return response.NewBadRequest(err)
-	}
-
-	if err := updateUserSlackToken(user, token.AccessToken); err != nil {
-		return response.NewBadRequest(err)
-	}
-
-	fmt.Printf("TOKEN IS: %+v", token)
-	return response.NewOK(nil)
-}
-
-func updateUserSlackToken(user *kodingmodels.User, m string) error {
-	selector := bson.M{"username": user.Name}
-	update := bson.M{"foreignAuth.slack.token": m}
-
-	return modelhelper.UpdateUser(selector, update)
-}
-
-type SlackRequest struct {
-	Token string
-}
-
-type SlackMessageRequest struct {
-	SlackRequest
-	Channel string
-	Text    string
-	Params  slack.PostMessageParameters
-}
-
-//
-// ListUsers & ListChannels & PostMessage handlers are used by client of koding
-//
-
-func (s *Slack) ListUsers(u *url.URL, h http.Header, req *SlackRequest, context *models.Context) (int, http.Header, interface{}, error) {
-	if !context.IsLoggedIn() {
-		return response.NewBadRequest(models.ErrNotLoggedIn)
-	}
-
-	token, err := getSlackToken(context.Client.Account)
-	if err != nil {
-		return response.NewBadRequest(err)
-	}
-
-	return response.HandleResultAndError(s.getUsers(token))
-}
-
-func (s *Slack) ListChannels(u *url.URL, h http.Header, req *SlackRequest, context *models.Context) (int, http.Header, interface{}, error) {
-	if !context.IsLoggedIn() {
-		return response.NewBadRequest(models.ErrNotLoggedIn)
-	}
-
-	token, err := getSlackToken(context.Client.Account)
-	if err != nil {
-		return response.NewBadRequest(err)
-	}
-
-	return response.HandleResultAndError(s.getChannels(token))
-}
-
-func (s *Slack) PostMessage(u *url.URL, h http.Header, req *SlackMessageRequest, context *models.Context) (int, http.Header, interface{}, error) {
-	if !context.IsLoggedIn() {
-		return response.NewBadRequest(models.ErrNotLoggedIn)
-	}
-
-	token, err := getSlackToken(context.Client.Account)
-	if err != nil {
-		return response.NewBadRequest(err)
-	}
-	req.Token = token
-	return response.HandleResultAndError(s.postMessage(req))
-}
-
-type SlackUser struct {
-	slack.User
-	LastActivity *time.Time `json:"lastActivity,omitempty"`
-}
-
-type SlackChannelsResponse struct {
-	Groups   []slack.Group   `json:"groups,omitempty"`
-	Channels []slack.Channel `json:"channels,omitempty"`
-}
-
-// getSlackToken fetches the user's slack token with user's accountID
-func getSlackToken(acc *models.Account) (string, error) {
-	var token string
-
-	user, err := modelhelper.GetUser(acc.Nick)
-	if err != nil {
-		return token, err
-	}
-
-	token = user.ForeignAuth.Slack.Token
-
-	if token == "" {
-		return token, models.ErrTokenIsNotFound
-	}
-
-	return token, nil
-}
-
-// getChannels send a request to the slack with user's token & gets the channels
-func (s *Slack) getChannels(token string) (*SlackChannelsResponse, error) {
-	api := slack.New(token)
-	var groups []slack.Group
-	var channels []slack.Channel
-	var gerr, cerr error
-	var wg sync.WaitGroup
-	wg.Add(2)
-
-	go func() {
-		groups, gerr = api.GetGroups(true)
-		wg.Done()
-	}()
-
-	go func() {
-		channels, cerr = api.GetChannels(true)
-		wg.Done()
-	}()
-
-	wg.Wait()
-
-	if gerr != nil {
-		return nil, gerr
-	}
-
-	if cerr != nil {
-		return nil, cerr
-	}
-
-	return &SlackChannelsResponse{
-		Groups:   groups,
-		Channels: channels,
-	}, nil
-}
-
-func (s *Slack) getUsers(token string) ([]SlackUser, error) {
-	api := slack.New(token)
-	allusers, err := api.GetUsers()
-	if err != nil {
-		return nil, err
-	}
-
-	var activeUsers []SlackUser
-	for _, u := range allusers {
-		// no need to add deleted users
-		if u.Deleted {
-			continue
-		}
-
-		// filter bots
-		if u.IsBot {
-			continue
-		}
-
-		p, err := api.GetUserPresence(u.ID)
-		if err != nil {
-			return nil, err
-		}
-
-		var lastActive *time.Time
-		if p != nil {
-			// set presence info
-			u.Presence = p.Presence
-
-			// LastActivity works inconsistently
-			// if u.Presence != slackAway {
-			// 	LastActiveTime := p.LastActivity.Time()
-			// 	lastActive = &LastActiveTime
-			// }
-		}
-
-		au := SlackUser{u, lastActive}
-
-		activeUsers = append(activeUsers, au)
-	}
-
-	return activeUsers, nil
-}
-
-func (s *Slack) postMessage(req *SlackMessageRequest) (string, error) {
-	api := slack.New(req.Token)
-	_, id, err := api.PostMessage(req.Channel, req.Text, req.Params)
-	return id, err
-}
